move content type enum to encrypted_store
