+++ xmtp_content_types/src/lib.rs
@@ -2,75 +2,8 @@ pub mod group_updated;
 pub mod membership_change;
 pub mod text;
 
-use diesel::{
-    backend::Backend,
-    deserialize::{self, FromSql, FromSqlRow},
-    expression::AsExpression,
-    serialize::{self, IsNull, Output, ToSql},
-    sql_types::Integer,
-    sqlite::Sqlite,
-};
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use xmtp_proto::xmtp::mls::message_contents::{ContentTypeId, EncodedContent};
-
-/// ContentType and their corresponding string representation
-/// are derived from the `ContentTypeId` enum in the xmtp-proto crate
-/// that each content type in this crate establishes for itself
-#[repr(i32)]
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, FromSqlRow, AsExpression)]
-#[diesel(sql_type = diesel::sql_types::Integer)]
-pub enum ContentType {
-    Unknown = 0,
-    Text = 1,
-    GroupMembershipChange = 2,
-    GroupUpdated = 3,
-}
-
-impl ContentType {
-    pub fn from_string(type_id: &str) -> Self {
-        match type_id {
-            text::TextCodec::TYPE_ID => Self::Text,
-            membership_change::GroupMembershipChangeCodec::TYPE_ID => Self::GroupMembershipChange,
-            group_updated::GroupUpdatedCodec::TYPE_ID => Self::GroupUpdated,
-            _ => Self::Unknown,
-        }
-    }
-
-    pub fn to_string(&self) -> &'static str {
-        match self {
-            Self::Unknown => "unknown",
-            Self::Text => text::TextCodec::TYPE_ID,
-            Self::GroupMembershipChange => membership_change::GroupMembershipChangeCodec::TYPE_ID,
-            Self::GroupUpdated => group_updated::GroupUpdatedCodec::TYPE_ID,
-        }
-    }
-}
-
-impl ToSql<Integer, Sqlite> for ContentType
-where
-    i32: ToSql<Integer, Sqlite>,
-{
-    fn to_sql<'b>(&'b self, out: &mut Output<'b, '_, Sqlite>) -> serialize::Result {
-        out.set_value(*self as i32);
-        Ok(IsNull::No)
-    }
-}
-
-impl FromSql<Integer, Sqlite> for ContentType
-where
-    i32: FromSql<Integer, Sqlite>,
-{
-    fn from_sql(bytes: <Sqlite as Backend>::RawValue<'_>) -> deserialize::Result<Self> {
-        match i32::from_sql(bytes)? {
-            0 => Ok(ContentType::Unknown),
-            1 => Ok(ContentType::Text),
-            2 => Ok(ContentType::GroupMembershipChange),
-            3 => Ok(ContentType::GroupUpdated),
-            x => Err(format!("Unrecognized variant {}", x).into()),
-        }
-    }
-}
 
 #[derive(Debug, Error)]
 pub enum CodecError {
