+++ xmtp_mls/src/storage/encrypted_store/group_message.rs
@@ -7,7 +7,7 @@ use diesel::{
     sql_types::Integer,
 };
 use serde::{Deserialize, Serialize};
-use xmtp_content_types::ContentType;
+use xmtp_content_types::{group_updated, membership_change, text};
 
 use super::{
     db_connection::DbConnection,
@@ -81,6 +81,61 @@ where
         match i32::from_sql(bytes)? {
             1 => Ok(GroupMessageKind::Application),
             2 => Ok(GroupMessageKind::MembershipChange),
+            x => Err(format!("Unrecognized variant {}", x).into()),
+        }
+    }
+}
+
+#[repr(i32)]
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, FromSqlRow, AsExpression)]
+#[diesel(sql_type = diesel::sql_types::Integer)]
+pub enum ContentType {
+    Unknown = 0,
+    Text = 1,
+    GroupMembershipChange = 2,
+    GroupUpdated = 3,
+}
+
+impl ContentType {
+    pub fn from_string(type_id: &str) -> Self {
+        match type_id {
+            text::TextCodec::TYPE_ID => Self::Text,
+            membership_change::GroupMembershipChangeCodec::TYPE_ID => Self::GroupMembershipChange,
+            group_updated::GroupUpdatedCodec::TYPE_ID => Self::GroupUpdated,
+            _ => Self::Unknown,
+        }
+    }
+
+    pub fn to_string(&self) -> &'static str {
+        match self {
+            Self::Unknown => "unknown",
+            Self::Text => text::TextCodec::TYPE_ID,
+            Self::GroupMembershipChange => membership_change::GroupMembershipChangeCodec::TYPE_ID,
+            Self::GroupUpdated => group_updated::GroupUpdatedCodec::TYPE_ID,
+        }
+    }
+}
+
+impl ToSql<Integer, Sqlite> for ContentType
+where
+    i32: ToSql<Integer, Sqlite>,
+{
+    fn to_sql<'b>(&'b self, out: &mut Output<'b, '_, Sqlite>) -> serialize::Result {
+        out.set_value(*self as i32);
+        Ok(IsNull::No)
+    }
+}
+
+impl FromSql<Integer, Sqlite> for ContentType
+where
+    i32: FromSql<Integer, Sqlite>,
+{
+    fn from_sql(bytes: <Sqlite as Backend>::RawValue<'_>) -> deserialize::Result<Self> {
+        match i32::from_sql(bytes)? {
+            0 => Ok(ContentType::Unknown),
+            1 => Ok(ContentType::Text),
+            2 => Ok(ContentType::GroupMembershipChange),
+            3 => Ok(ContentType::GroupUpdated),
             x => Err(format!("Unrecognized variant {}", x).into()),
         }
     }
